Revert "Merge branch 'master' of github.com:omanamos/kinect-nao"
This reverts commit 7cad0d5b4c5efc8b56331c8f2985b7650dad5d37, reversing
changes made to c8b7098d47c919dff789c62f0a56f4e901080fd1.

+++ DataStore/NaoSkeleton/NaoPosition.cs
@@ -7,13 +7,13 @@ namespace DataStore
 {
     public class NaoPosition
     {
-        public float X { get; private set; }
-        public float Y { get; private set; }
-        public float Z { get; private set; }
+        public double X { get; private set; }
+        public double Y { get; private set; }
+        public double Z { get; private set; }
 
         private NaoPosition() { }
 
-        public NaoPosition(float x, float y, float z)
+        public NaoPosition(double x, double y, double z)
         {
             this.X = x;
             this.Y = y;
